add base parser for reusing code for for-loop parser
